Resolution multiplier and support for play mode
Kinda, you have to drag the prefab into the scene for the script to work in play mode. Also, there are two separate scripts, one for play mode and one for the editor. To change the resolution in play mode you can just adjust the value in the public resolutionMultiplier field. I couldn't find a way to expose a parameter in an editor script so you will have to change it manually there.